move secondary attack to its own manager
